[Refactor] Extract upgrade fixtures from tests/conftest.py
Move upgrade-domain fixtures out of the monolithic tests/conftest.py
into two domain modules:
- tests/fixtures/upgrade/config.py
- tests/fixtures/upgrade/resources.py

Move determine_upgrade_stream() helper to utilities/operator.py with
type hints and a Google-style docstring.

Register new modules in pytest_plugins (root conftest.py).
