+++ conftest.py
@@ -79,6 +79,8 @@ pytest_plugins = [
     "tests.fixtures.network.cluster",
     "tests.fixtures.images.validation_os_images",
     "tests.fixtures.network.multiarch",
+    "tests.fixtures.upgrade.config",
+    "tests.fixtures.upgrade.resources",
 ]
 
 LOGGER = logging.getLogger(__name__)
